+++ .env.example
@@ -8,7 +8,9 @@ API_VERSION=1
 
 # JWT AUTH
 JWT_SECRET_KEY=
+JWT_REFRESH_SECRET=
 JWT_EXPIRATION_TIME=3600
+JWT_REFRESH_EXPIRATION_TIME=7d
 
 # DATABASE
 DB_TYPE=postgres
