+++ src/modules/auth/events/email.event.ts
