+++ src/modules/chapter-management/chapter-management.controller.ts
@@ -1,10 +1,11 @@
 import { Crud, CrudController } from '@nestjsx/crud';
-import { Controller } from '@nestjs/common';
+import { Controller, UseGuards } from '@nestjs/common';
 import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
 import { Roles } from '../common/decorator/roles.decorator';
 import { AppRoles } from '../common/enum/roles.enum';
 import { ChapterEntity } from './entity/chapter.entity';
 import { ChapterManagementService } from './chapter-management.service';
+import { JwtAuthGuard } from '../common/guard/jwt-guard';
 
 @Crud({
   model: {
@@ -17,6 +18,7 @@ import { ChapterManagementService } from './chapter-management.service';
 })
 @ApiTags('Chapters Management')
 @ApiBearerAuth()
+@UseGuards(JwtAuthGuard)
 @Roles(AppRoles.ADMINS)
 export class ChapterManagementController
   implements CrudController<ChapterEntity>
