+++ src/modules/common/interface/request-with-user.interface.ts
@@ -1,0 +1,8 @@
+import { Request } from 'express';
+import { UserEntity } from '../../user/entity/user.entity';
+
+interface RequestWithUser extends Request {
+  user: UserEntity;
+}
+
+export default RequestWithUser;
