+++ src/modules/common/strategy/refresh-token.strategy.ts
@@ -1,0 +1,26 @@
+import { ExtractJwt, Strategy } from 'passport-jwt';
+import { PassportStrategy } from '@nestjs/passport';
+import { Injectable } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
+// import { UsersService } from 'src/modules/user/user.service';
+import { Request } from 'express';
+
+@Injectable()
+export class RefreshTokenStrategy extends PassportStrategy(
+  Strategy,
+  'jwt-refresh',
+) {
+  constructor(readonly configService: ConfigService) {
+    super({
+      jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
+      ignoreExpiration: false,
+      passReqToCallback: true,
+      secretOrKey: configService.get('JWT_REFRESH_SECRET'),
+    });
+  }
+
+  validate(req: Request, payload: any) {
+    const refreshToken = req.get('Authorization').replace('Bearer', '').trim();
+    return { ...payload, refreshToken };
+  }
+}
