--- src/modules/user-auth/user-auth.module.ts
@@ -1,31 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { PassportModule } from '@nestjs/passport';
-import { UserAuthService } from './user-auth.service';
-import { UserAuthController } from './user-auth.controller';
-import { ConfigModule, ConfigService } from '@nestjs/config';
-import { JwtModule } from '@nestjs/jwt';
-import EmailEvent from './events/email.event';
-import { UserEntity } from '../user/entity/user.entity';
-import { ForgotEntity } from './entity/forgot.entity';
-@Module({
-  imports: [
-    ConfigModule,
-    TypeOrmModule.forFeature([UserEntity, ForgotEntity]),
-    PassportModule.register({ defaultStrategy: 'local' }),
-    JwtModule.registerAsync({
-      imports: [ConfigModule],
-      inject: [ConfigService],
-      useFactory: async (configService: ConfigService) => ({
-        secret: configService.get('JWT_SECRET'),
-        signOptions: {
-          expiresIn: `${configService.get('JWT_EXPIRATION_TIME')}s`,
-        },
-      }),
-    }),
-  ],
-  exports: [UserAuthService],
-  controllers: [UserAuthController],
-  providers: [UserAuthService, EmailEvent],
-})
-export class UserAuthModule {}
