+++ src/modules/user/user.service.ts
@@ -2,22 +2,16 @@ import {
   BadRequestException,
   Injectable,
   NotAcceptableException,
-  UnauthorizedException,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { UserEntity } from './entity/user.entity';
-import { Hash } from '../../utils/Hash';
 import {
   IPaginationOptions,
   paginate,
   Pagination,
 } from 'nestjs-typeorm-paginate';
-import { ResetPayload } from '../auth/payloads/reset.payload';
 import { UpdatePayload } from './payloads/update.payload';
-import { RegisterPayload } from '../auth/payloads/register.payload';
-import { PurchaseEntity } from '../purchase/entity/purchase.entity';
-import { PurchaseEnum } from '../common/enum/purchase.enum';
 import { TypeOrmCrudService } from '@nestjsx/crud-typeorm';
 
 @Injectable()
@@ -25,8 +19,6 @@ export class UsersService extends TypeOrmCrudService<UserEntity> {
   constructor(
     @InjectRepository(UserEntity)
     private readonly userRepository: Repository<UserEntity>,
-    @InjectRepository(PurchaseEntity)
-    private readonly purchaseRepository: Repository<PurchaseEntity>,
   ) {
     super(userRepository);
   }
@@ -67,44 +59,44 @@ export class UsersService extends TypeOrmCrudService<UserEntity> {
     return paginate<UserEntity>(queryBuilder, options);
   }
 
-  async getCourse(
-    id: string,
-    options: IPaginationOptions,
-  ): Promise<Pagination<PurchaseEntity>> {
-    const queryBuilder = this.purchaseRepository.createQueryBuilder('purchase');
-    queryBuilder.andWhere('purchase.byUserId =:id', {
-      id: id,
-    });
-    queryBuilder.andWhere('purchase.status =:status', {
-      status: PurchaseEnum.VERIFIED,
-    });
-    queryBuilder.leftJoinAndSelect('purchase.courseId', 'purchasedCourses');
-    return paginate<PurchaseEntity>(queryBuilder, options);
-  }
+  // async getCourse(
+  //   id: string,
+  //   options: IPaginationOptions,
+  // ): Promise<Pagination<PurchaseEntity>> {
+  //   const queryBuilder = this.purchaseRepository.createQueryBuilder('purchase');
+  //   queryBuilder.andWhere('purchase.byUserId =:id', {
+  //     id: id,
+  //   });
+  //   queryBuilder.andWhere('purchase.status =:status', {
+  //     status: PurchaseEnum.VERIFIED,
+  //   });
+  //   queryBuilder.leftJoinAndSelect('purchase.courseId', 'purchasedCourses');
+  //   return paginate<PurchaseEntity>(queryBuilder, options);
+  // }
 
-  async changPassword(payload: ResetPayload): Promise<any> {
-    const user = await this.getByUsername(payload.username);
-    if (!user || !Hash.compare(payload.currentPassword, user.password)) {
-      throw new UnauthorizedException('Invalid credentials!');
-    }
-    await this.userRepository
-      .createQueryBuilder('users')
-      .update(UserEntity)
-      .set({ password: payload.newPassword })
-      .where('username =:username', { username: payload.username })
-      .execute();
-    return user;
-  }
+  // async changPassword(payload: ResetPayload): Promise<any> {
+  //   const user = await this.getByUsername(payload.username);
+  //   if (!user || !Hash.compare(payload.currentPassword, user.password)) {
+  //     throw new UnauthorizedException('Invalid credentials!');
+  //   }
+  //   await this.userRepository
+  //     .createQueryBuilder('users')
+  //     .update(UserEntity)
+  //     .set({ password: payload.newPassword })
+  //     .where('username =:username', { username: payload.username })
+  //     .execute();
+  //   return user;
+  // }
 
-  async create(payload: RegisterPayload) {
-    const user = await this.getByUsername(payload.username);
-    if (user) {
-      throw new NotAcceptableException(
-        'Admin with provided username already created.',
-      );
-    }
-    return await this.userRepository.save(this.userRepository.create(payload));
-  }
+  // async create(payload: RegisterPayload) {
+  //   const user = await this.getByUsername(payload.username);
+  //   if (user) {
+  //     throw new NotAcceptableException(
+  //       'Admin with provided username already created.',
+  //     );
+  //   }
+  //   return await this.userRepository.save(this.userRepository.create(payload));
+  // }
 
   async delete(id: string): Promise<any> {
     const user = await this.userRepository.findOne(id);
